requirements.txt
flask
requests
kafka-python 
